Refactoring: CSS > SASS #1
