+++ src/components/ExperienceBar.tsx
@@ -2,20 +2,20 @@ import { useContext } from 'react'
 import { ChallengesContext } from '../contexts/ChallengesContext'
 
 // CSS - module.css //
-import styles from '../styles/components/ExperienceBar.module.css'
+import styles from '../styles/components/ExperienceBar.module.scss'
 
 export function ExperienceBar() {
-
     const { currentExperience, experienceToNextLevel } = useContext(ChallengesContext)
 
     const percentToNextLevel = Math.round(currentExperience * 100) / experienceToNextLevel
+
     return (
         <header className={styles.experienceBar}>
             <span>
                 0 XP
             </span>
             <div>
-                <div style={{ width: `${percentToNextLevel}%` }} />
+                <div style={{ width: `${percentToNextLevel}%` }} className={styles.progressBar}/>
 
                 <span className={styles.currentExperience} style={{ left: `${percentToNextLevel}%` }}>
                     {currentExperience} XP
