--- src/styles/components/Countdown.module.css
@@ -1,91 +1,0 @@
-.countdownContainer{
-    display: flex;
-    align-items: center;
-
-    font-family: Rajdhani;
-    font-weight: 600;
-    color: var(--title);
-}
-
-.countdownContainer > div{
-    flex: 1;
-
-    display: flex;
-    align-items: center;
-    justify-content: space-evenly;
-    
-    background: var(--white);
-    box-shadow: 0 0 60px rgba(0, 0, 0, 0.05);
-    border-radius: 5px;
-    font-size:  8.5rem;
-    text-align: center;
-}
-
-.countdownContainer > div span{
-    flex: 1;
-}
-
-.countdownContainer > div span:first-child{
-    border-right: 1px solid #f0f1f3;
-}
-
-.countdownContainer > div span:last-child{
-    border-left: 1px solid #f0f1f3;
-}
-
-.countdownContainer > span{
-    font-size: 6.25rem;
-    margin: 0 0.5rem;
-}
-
-.startCountdownButton{
-    width: 100%;
-    height: 5rem;
-    margin-top: 2rem;
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 0;
-    border-radius: 5px;
-    background: var(--blue);
-    color: var(--white);
-    outline: none;
-    
-    font-size: 1.25rem;
-    font-weight: 600;
-    transition: 0.2s background;
-}
-
-.startCountdownButton:not(:disabled):hover{
-    background: var(--blue-dark);
-}
-
-.startCountdownButtonActive{
-    background: var(--white);
-    color: var(--text);
-
-    transition: 0.2 background;
-}
-
-.startCountdownButtonActive:not(:disabled):hover{
-    background: var(--red);
-    color: var(--white);
-}
-
-.startCountdownButtonActive:disabled{
-    cursor: not-allowed;
-    background: var(--white);
-    color: var(--text);
-    border-bottom: solid 4px var(--green)
-}
-
-.startCountdownButtonActive:disabled svg{
-    color: var(--green);
-}
-
-.startCountdownButton svg{
-    width: 1.15rem;
-    height: 1.15rem;
-    margin-left: 0.8rem;
-}
