--- src/styles/components/ExperienceBar.module.css
@@ -1,30 +1,0 @@
-.experienceBar {
-    display: flex;
-    align-items: center;
-}
-
-.experienceBar span{
-    font-size:  1rem;
-}
-
-.experienceBar > div{
-    flex: 1;
-    height: 4px;
-    border-radius: 4px;
-    background: var(--gray-line);
-    margin: 0 1.5rem;
-    position: relative;
-}
-
-.experienceBar > div > div {
-    height: 4px;
-    border-radius: 4px;
-    background: var(--green);
-
-}
-
-span.currentExperience{
-    position: absolute;
-    top: 12px;
-    transform: translateX(-50%); /* Hackzinho */
-}
