+++ src/styles/components/ExperienceBar.module.scss
@@ -1,0 +1,29 @@
+.experienceBar {
+    display: flex;
+    align-items: center;
+
+    span {
+        font-size: 1rem;
+    }
+
+    & > div {
+        flex: 1;
+        height: 4px;
+        border-radius: 4px;
+        background: var(--gray-line);
+        margin: 0 1.5rem;
+        position: relative;
+    }
+}
+
+.progressBar {
+    height: 4px;
+    border-radius: 4px;
+    background: var(--green);
+}
+
+.currentExperience {
+    position: absolute;
+    top: 12px;
+    transform: translateX(-50%); /* Hackzinho */
+}
