--- src/styles/components/Perfil.module.css
@@ -1,30 +1,0 @@
-.perfilContainer{
-
-    display: flex;
-    align-items: center;
-}
-
-.perfilContainer > img {
-    width: 5.5rem;
-    height: 5.5rem;
-    border-radius: 50%;
-}
-
-.perfilContainer div {
-    margin-left: 1.5rem;
-}
-
-.perfilContainer div strong{
-    font-size: 1.5rem;
-    font-weight: 600;
-    color: var(--title);
-}
-
-.perfilContainer div p{
-    font-size: 1rem;
-    margin-top:  0.5rem;
-}
-
-.perfilContainer div p img {
-    margin-right: 0.5rem;
-}
