+++ src/styles/components/Perfil.module.scss
@@ -1,0 +1,29 @@
+.perfilContainer {
+    display: flex;
+    align-items: center;
+
+    > img {
+        width: 5.5rem;
+        height: 5.5rem;
+        border-radius: 50%;
+    }
+
+    div {
+        margin-left: 1.5rem;
+
+        strong {
+            font-size: 1.5rem;
+            font-weight: 600;
+            color: var(--title);
+        }
+
+        p {
+            font-size: 1rem;
+            margin-top: 0.5rem;
+
+            img {
+                margin-right: 0.5rem;
+            }
+        }
+    }
+}
