--- src/styles/pages/Home.module.css
@@ -1,20 +1,0 @@
-.container{
-    height: 100vh;
-    max-width: 992px;
-    margin: 0 auto;
-    padding: 2.5rem 2rem;
-
-    display: flex;
-    flex-direction: column;
-}
-
-.container > section {
-    flex: 1;
-
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 6.25rem;
-
-    align-content: center;
-    
-}
